Safari-611.1.10/LayoutTests/animations/leak-document-with-css-animation-expected.txt

This test asserts that Document doesn't leak when a CSS Animation is created.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The iframe has finished loading.
The iframe has been destroyed.
PASS internals.numberOfLiveDocuments() is numberOfLiveDocumentsAfterIframeLoaded - 1

PASS successfullyParsed is true

TEST COMPLETE
